Fix multiple bugs and add Docker support for modern compilation
- Fix pointer truncation on 64-bit: change player->work from int to
  intptr_t, update all casts and callback signatures (change.c, frob.c,
  sendsys.c, mud.c, comm.c, climate.c)
- Fix use-after-free on first login: check getuaf() return value (mud.c)
- Fix mailer crash on NULL file pointer in quit_mailer() (mailer.c)
- Fix missing break in switch case 'u' causing fall-through (main.c)
- Fix file descriptor leak in syslogcom() (wizard.c)
- Fix NULL pointer dereference in bootstrap when fopen fails (bootstrap.c)
- Fix pclose/fclose mismatch in convert utility (convert.c)
- Fix unsafe pmaxmagic macro missing parentheses (mudmacros.h)
- Fix test.sh syntax error: missing space before ] (test.sh)
- Replace deprecated sys_errlist with strerror() (bootstrap.c, log.c, timing.c)
- Fix socket binding to INADDR_ANY for container compatibility (s_socket.c)
- Add -fcommon flag for GCC 10+ compatibility (Makefile)
- Add HDIR include path definition to Makefile
- Add #include <stdint.h> for intptr_t support (kernel.h, sendsys.h)
- Add new Dockerfile for Ubuntu 22.04 with proper build steps
- Add setlevel utility for setting player levels
- Fix wizlist pointer-to-int truncation (wizlist.c)

+++ src/mudtypes.h
@@ -306,7 +306,7 @@ typedef struct _a_player {
   char		*sock_buf_end;
   void		*writer;	/* Current writer */
   int		no_logins;	/* No of failed passwd */
-  int		work;		/* General work area for use by code */
+  intptr_t	work;		/* General work area for use by code */
   char		work2[64];
   Boolean	no_echo;	/* True if telnet should turn off echo */
   Boolean	isawiz;		/* Is this player a system's wizard? */
